QueryCache hits and misses must be long instead of int, otherwise may run over
git-svn-id: https://svn.exedio.com/cope/trunk/persistence@8139 e7d4fc99-c606-0410-b9bf-843393a9eab7

+++ runtime/src/com/exedio/cope/Model.java
@@ -783,7 +783,7 @@ public final class Model
 		return getItemCacheInfo();
 	}
 	
-	public int[] getQueryCacheInfo()
+	public long[] getQueryCacheInfo()
 	{
 		return getQueryCache().getQueryInfo();
 	}
@@ -792,7 +792,7 @@ public final class Model
 	 * @deprecated renamed to {@link #getQueryCacheInfo()}.
 	 */
 	@Deprecated
-	public int[] getCacheQueryInfo()
+	public long[] getCacheQueryInfo()
 	{
 		return getQueryCacheInfo();
 	}
